deprecated調整
